% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMMI_index.R
\name{AMMI_index}
\alias{AMMI_index}
\title{AMMI Stability Function}
\usage{
AMMI_index(data)
}
\arguments{
\item{data}{A numeric data frame with the columns in the following order: genotypes,
environments, replications and the response variable.}
}
\description{
Performs AMMI stability analysis
}
