% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gabriel.R
\name{Gabriel.Calinski}
\alias{Gabriel.Calinski}
\title{Gabriel imputation}
\usage{
Gabriel.Calinski(X)
}
\arguments{
\item{X}{a data frame or matrix that contains the genotypes in the rows and the
environments in the columns when there are no replications of the experiment.
#'@references Arciniegas-Alarcón S., García-Peña M., Dias C.T.S., Krzanowski W.J.. (2010).
\emph{An alternative methodology for imputing missing data in trials with genotype-by-environment interaction}.
Biometrical Letters, 47, 1–14.}
}
\description{
Gabriel imputation
}
