% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGE_Plot.R
\name{GGEPlot}
\alias{GGEPlot}
\title{GGE biplots with \pkg{ggplot2}}
\usage{
GGEPlot(
  GGEModel,
  type = "Biplot",
  d1 = 1,
  d2 = 2,
  selectedE = NA,
  selectedG = NA,
  selectedG1 = NA,
  selectedG2 = NA,
  colGen = "gray47",
  colEnv = "darkred",
  colSegment = "gray30",
  colHull = "gray30",
  sizeGen = 4,
  sizeEnv = 4,
  largeSize = 4.5,
  axis_expand = 1.2,
  axislabels = TRUE,
  axes = TRUE,
  limits = TRUE,
  titles = TRUE,
  footnote = TRUE
)
}
\arguments{
\item{GGEModel}{An object of class \code{GGEModel} or \code{gge}}

\item{type}{type of biplot to produce. \enumerate{\item Basic biplot. \item
Examine environment. See \code{\link[GGEBiplots]{ExamineEnv}} \item Examine
genotype. See \code{\link[GGEBiplots]{ExamineGen}} \item Relationship among
environments. See \code{\link[GGEBiplots]{EnvRelationship}} \item Compare
two genotypes. See \code{\link[GGEBiplots]{CompareGens}} \item Which won
where/what. See \code{\link[GGEBiplots]{WhichWon}} \item Discrimination vs.
representativeness. See \code{\link[GGEBiplots]{DiscRep}} \item Ranking
environments. See \code{\link[GGEBiplots]{RankEnv}} \item Mean vs.
stability. See \code{\link[GGEBiplots]{MeanStability}} \item Ranking
gentoypes See \code{\link[GGEBiplots]{RankGen}}}}

\item{d1}{PCA component to plot on x axis. Defaults to 1}

\item{d2}{PCA component to plot on y axis. Defaults to 2}

\item{selectedE}{name of the environment to examine when type=2. Must be a 
string which matches an environment label}

\item{selectedG}{name of the genotype to examine when type=3. Must be a string
which matches a genotype label}

\item{selectedG1}{name of a genotype to compare when type=5. Must be a string 
which matches a genotype label}

\item{selectedG2}{name of a genotype to compare when type=5. Must be a string 
which matches a genotype label and not equal to selectedG1}

\item{colGen}{colour for genotype attributes on biplot. Defaults to 
"forestgreen"}

\item{colEnv}{colour for environment attributes on biplot. Defaults to "blue"}

\item{colSegment}{colour for segment or circle lines. Defaults to "red"}

\item{colHull}{colour for hull when type=6. Defaults to "black"}

\item{sizeGen}{text size for genotype labels. Defaults to 4}

\item{sizeEnv}{text size for environment labels. Defaults to 4}

\item{largeSize}{text size to use for larger labels where type=5, used for the
two selected genotypes, and where type=6, used for the outermost genotypes.
Defaults to 4.5}

\item{axis_expand}{multiplication factor to expand the axis limits by to 
enable fitting of labels. Defaults to 1.2}

\item{axislabels}{logical. If TRUE then include automatically generated labels
for axes}

\item{axes}{logical. If TRUE then include x and y axes going through the 
origin}

\item{limits}{logical. If TRUE then automatically rescale axes}

\item{titles}{logical. If TRUE then include automatically generated titles}

\item{footnote}{logical. If TRUE then include automatically generated footbote}
}
\value{
A biplot of class \code{ggplot}
}
\description{
The GGE biplot allows a visual examination of the relationships
 between test environments, genotypes, and genotype interactions by
 environment. The GGE biplot is an object of class 'ggplot' obtained from an
 object produced by the \code{\link[geneticae]{GGEmodel}} function. It is
 possible to personalize the graphic so that the stylistic attributes are
 liked by the user. This function is a modification of the
 \code{\link[GGEBiplots]{GGEPlot}} package of the GGEBiplots.
}
\examples{
 library(geneticae)

 # Data without replication
 data(yan.winterwheat)
 GGE1 <- GGEmodel(yan.winterwheat, centering = "tester")
 GGEPlot(GGE1)

 # Data with replication
 data(plrv)
 GGE2 <- GGEmodel(plrv,genotype="Genotype",environment="Locality",
 response="Yield", rep="Rep", centering = "tester")
 GGEPlot(GGE2)

}
\references{
Yan W, Kang M (2003). \emph{GGE Biplot Analysis: A Graphical Tool
 for Breeders, Geneticists, and Agronomists}. CRC Press.

Sam Dumble (2017). GGEBiplots: GGE Biplots with 'ggplot2'. R
 package version 0.1.1. \url{https://CRAN.R-project.org/package=GGEBiplots}
}
\keyword{Biplot}
\keyword{GGE}
