% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{plrv}
\alias{plrv}
\title{Clones from the PLRV population}
\format{
A data frame with 504 observations on the following 6 variables

 Genotype a factor with levels 102.18 104.22 121.31 141.28 157.26 163.9
 221.19 233.11 235.6 241.2 255.7 314.12 317.6 319.20 320.16 342.15 346.2
 351.26 364.21 402.7 405.2 406.12 427.7 450.3 506.2 Canchan Desiree Unica.

 Locality a factor with levels Ayacucho (Ayac), La Molina 02 (LM-02), San
 Ramon 02 (SR-02), Huancayo (Hyo-02), La Molina 03 (LM-03), San Ramon 03
 (SR-03).

 Rep a numeric vector

 WeightPlant, WeightPlot and Yield are numeric vectors
}
\usage{
data(plrv)
}
\description{
Resistance study to PLRV (\emph{Patato Leaf Roll Virus}) that
 causes leaf curl. 28 genotypes were experimented at 6 locations in Peru.
 Each clone was evaluated three times in each environment, and the yield,
 plant weight and plot were counted for each one. This dataset was obtained
 from the Agricolae Package.
}
\examples{
library(geneticae)
data(plrv)
str(plrv)

}
\references{
Felipe de Mendiburu (2020). agricolae: Statistical Procedures for
 Agricultural Research. R package version 1.3-2.
 \url{https://CRAN.R-project.org/package=agricolae}
}
\keyword{datasets}
