% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rAMMIPlot.R
\name{rAMMIPlot}
\alias{rAMMIPlot}
\title{AMMI Biplots with ggplot2}
\usage{
rAMMIPlot(
  model_res,
  colGen = "gray47",
  colEnv = "darkred",
  sizeGen = 6,
  sizeEnv = 6,
  titles = TRUE,
  footnote = TRUE,
  axis_expand = 1.2,
  limits = TRUE,
  axes = TRUE,
  axislabels = TRUE
)
}
\arguments{
\item{model_res}{an object of class 'AMMI' from AMMIModel.}

\item{colGen}{genotype colour. Defaults to "gray47".}

\item{colEnv}{environment colour. Defaults to "darkred".}

\item{sizeGen}{genotype text size.}

\item{sizeEnv}{environment text size.}

\item{titles}{logical, show plot title.}

\item{footnote}{logical, show footnote with explained variance.}

\item{axis_expand}{expansion factor for axis limits.}

\item{limits}{logical. If `TRUE` axes are automatically rescaled. Defaults to
`TRUE`.}

\item{axes}{logical, if this argument is `TRUE` axes passing through the
origin are drawn. Defaults to `TRUE`.}

\item{axislabels}{logical, if this argument is `TRUE` labels axes are included.
Defaults to `TRUE`}
}
\description{
Produces a biplot for objects of class 'AMMI'.
}
