% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{stability}
\alias{stability}
\title{Stability Function}
\usage{
stability(data, type = "fox")
}
\arguments{
\item{data}{A numeric data.frame with fenotypic means values of MET, and genotypes and environments by rows and columns, respectively.}

\item{type}{type of biplot to produce. fox, kang,nahu, thsu, AMMIindex}
}
\description{
Performs a stability analysis
}
