% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W_GabrielEigen.R
\name{WGabriel}
\alias{WGabriel}
\title{Weighted GabrielEigen imputation method}
\usage{
WGabriel(DBmiss, Winf, Wsup)
}
\arguments{
\item{DBmiss}{a data frame or matrix that contains the genotypes in the rows
and the environments in the columns when there are no replications of the
experiment.}

\item{Winf}{inferior weight}

\item{Wsup}{superior weight}
}
\value{
A list containing: \itemize{\item Peso: weight that provides the best
 predictive difference; \item NumeroIterWGabriel: the final number of
 iterations; \item CritConvergWGabriel: the maximum change of the estimated
 values for missing cells in the last step of iteration (the precision of
 convergence);\item GabrielWImput: the imputed matrix (filled in with the
 missing values estimated by the Weighted GabrielEigen procedure).}
}
\description{
Weighted GabrielEigen imputation method
}
\references{
Arciniegas-Alarcón S., García-Peña M., Krzanowski W.J., Dias
 C.T.S. (2014). An alternative methodology for imputing missing data in
 trials with genotype-byenvironment interaction: some new aspects.
 Biometrical Letters 51, 75-88.
}
\keyword{internal}
