% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{yan.winterwheat}
\alias{yan.winterwheat}
\title{Winter wheat varieties from Ontario}
\format{
A data frame with 162 observations and 3 variables (genotype,
  environment and yield)
}
\usage{
data(yan.winterwheat)
}
\description{
Yield (t/ha) of 18 winter wheat varieties grown at 9 environments in
  Ontario in 1993.
}
\examples{
library(geneticae)
data(yan.winterwheat)
str(yan.winterwheat)
}
\references{
Kevin Wright (2018). agridat: Agricultural Datasets. R package
  version 1.16.\url{https://CRAN.R-project.org/package=agridat}
}
\keyword{datasets}
