% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GabrielEigen.R
\name{Gabriel.Calinski}
\alias{Gabriel.Calinski}
\title{GabrielEigen imputation method}
\usage{
Gabriel.Calinski(X)
}
\arguments{
\item{X}{a data frame or matrix that contains genotypes in rows and
environments in columns when there are no replications of the
experiment.}
}
\value{
A list containing: \itemize{ \item
  NumeroIterGabriel: final number of iterations; \item
  CritConvergGabriel: maximum change of the estimated values for missing
  cells in the last step of iteration (precision of convergence); \item
  GabrielImput: imputed matrix (filled in with missing values
  estimated by the GabrielEigein procedure).}
}
\description{
GabrielEigen imputation method
}
\references{
Arciniegas-Alarcón S., García-Peña M., Dias C.T.S., Krzanowski
  W.J. (2010). \emph{An alternative methodology for imputing missing data in
  trials with genotype-by-environment interaction}. Biometrical Letters 47,
  1–14.
}
\keyword{internal}
