% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGE_Model.R
\name{GGEmodel}
\alias{GGEmodel}
\title{Site regression model}
\usage{
GGEmodel(
  Data,
  genotype = "gen",
  environment = "env",
  response = "yield",
  rep = NULL,
  centering = "tester",
  scaling = "none",
  SVP = "column"
)
}
\arguments{
\item{Data}{a dataframe with genotypes, environments, repetitions (if any) and
the phenotypic trait of interest. There is no restriction on the order in
which these variables should be presented in the dataframe, and also other
variables that will not be used in the analysis can be included.}

\item{genotype}{column name containing genotypes.}

\item{environment}{column name containing environments.}

\item{response}{column name containing phenotypic trait.}

\item{rep}{column name containing replications. If this argument is NULL,
there is no replication available on the data.}

\item{centering}{centering method. Either "tester" for tester centered (G+GE),
"global" for global centered (E+G+GE), "double" for double centred (GE) or
"none" for no centering. If a centering method is not used, the
\code{\link[geneticae]{GGEPlot}} function can not be used.}

\item{scaling}{scaling method. Either "sd" for standard deviation or "none"
for no scaling.}

\item{SVP}{method for singular value partitioning. Either "row","column",
"dual" or "symmetrical".}
}
\value{
A list of class \code{GGE_Model} containing:
 \item{coordgenotype}{plotting coordinates for genotypes from all components}
 \item{coordenviroment}{plotting coordinates for environments from all
 components} \item{eigenvalues}{vector of eigenvalues from each component}
 \item{vartotal}{overall variance} \item{varexpl}{percentage of variance
 explained by each component} \item{labelgen}{genotype names}
 \item{labelenv}{environment names} \item{axes}{axis labels}
 \item{Data}{scaled and centered input data} \item{centering}{name of
 centering method} \item{scaling}{name of scaling method} \item{SVP}{name of
 SVP method}
}
\description{
The site regression model also named genotype plus
 genotype-by-environment (GGE) model is powerful a tools for effective
 analysis and interpretation of multienvironment data structure in breeding
 programs.This function is a modification of
 \code{\link[GGEBiplots]{GGEModel}} of GGEBiplots package, where the dataset
 contains genotype by environment means with the genotypes in  rows and
 environments in columns. By contrast, this function allows a less
 restrictive format of the data set. Repetitions and other variables theat
 will not be used in the analysis may be present in the dataset.
}
\examples{

 library(geneticae)
 # Data without replication
 GGE1 <- GGEmodel(yan.winterwheat, genotype="gen",environment="env", response="yield",
 centering = "tester")

 # Data with replication
 data(plrv)
 GGE2 <- GGEmodel(plrv, genotype="Genotype",environment="Locality",
 response="Yield", rep="Rep", centering = "tester")

}
\references{
Yan W, Kang M (2003). \emph{GGE Biplot Analysis: A Graphical Tool
 for Breeders, Geneticists, and Agronomists}. CRC Press.

Yan W, Kang M (2002). \emph{Singular-Value Partitioning in Biplot
 Analysis of Multienvironment Trial Data}. Agronomy Journal, 94, 990-996.
 \url{http://dx.doi.org/10.2134/agronj2002.0990}
}
