% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{imputation}
\alias{imputation}
\title{Produces genotype plus genotype-by-environment model from a 2-way table of
means}
\usage{
imputation(Data, rep = FALSE, type = "EM-AMMI")
}
\arguments{
\item{Data}{a data frame or matrix that contains the genotypes in the rows and the
environments in the columns when there are no replications of the experiment. In case
of replications, the data frame must have the columns in the following order: genotypes,
environments, replications and the response variable.}

\item{rep}{logical. If TRUE the genotype by environment means is calculated.}

\item{type}{imputation method. Either "EM-AMMI", "EM-SVD", "Gabriel","WGabriel"}
}
\value{
matrix with imputed data
}
\description{
Produces genotype plus genotype-by-environment model from a 2-way table of
means
}
\examples{
  library(geneticae)
  data(Ontario)
  imputation(Ontario, rep=FALSE, type="EM-SVD")
}
