% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{plrv}
\alias{plrv}
\title{Clones from the PLRV population}
\format{
Data frame with 504 observations and 6 variables (genotype,
  locality, repetition, weightPlant, weightPlot and yield).
}
\usage{
data(plrv)
}
\description{
Yield, plant weight and plot of 28 genotypes in Peru.
}
\examples{
library(geneticae)
data(plrv)
str(plrv)

}
\references{
Felipe de Mendiburu (2020). agricolae: Statistical Procedures for
 Agricultural Research. R package version 1.3-2.
 \url{https://CRAN.R-project.org/package=agricolae}
}
\keyword{datasets}
