% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{plrv}
\alias{plrv}
\title{Data clones from the PLRV population}
\format{A data frame with 504 observations on the following 6 variables.
Genotype a factor with levels 102.18 104.22 121.31 141.28 157.26 163.9 221.19 233.11
235.6 241.2 255.7 314.12 317.6 319.20 320.16 342.15 346.2 351.26 364.21 402.7
405.2 406.12 427.7 450.3 506.2 Canchan Desiree Unica
Locality a factor with levels Ayac Hyo-02 LM-02 LM-03 SR-02 SR-03
Rep a numeric vector
WeightPlant a numeric vector
WeightPlot a numeric vector
Yield a numeric vector}
\usage{
data(plrv)
}
\description{
Six environments: Ayacucho, La Molina 02, San Ramon 02, Huancayo,
La Molina 03, San Ramon 03. This data.frame was obtained from agricolae Package.
}
\examples{
library(geneticae)
data(plrv)
str(plrv)

}
\references{
AAAAA
}
\keyword{datasets}
