% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-AMMI.R
\name{rAMMI}
\alias{rAMMI}
\title{AMMI biplots with \pkg{ggplot2}}
\usage{
rAMMI(x, Ncomp = 2, type = "AMMI", rep = FALSE, colGen = "gray47",
  colEnv = "darkred", colSegment = "gray30", colHull = "gray30",
  sizeGen = 4, sizeEnv = 4, largeSize = 4.5, titles = TRUE,
  footnote = TRUE)
}
\arguments{
\item{x}{a data frame or matrix that contains the genotypes in the rows and the
environments in the columns when there are no replications of the experiment. In case
of replications, the data frame must have the columns in the following order: genotypes,
environments, replications and the response variable.}

\item{Ncomp}{number of principal components that will be used in the analysis}

\item{type}{method. Either "AMMI", "rAMMI", "hAMMI", "gAMMI", "lAMMI" or "ppAMMI".}

\item{rep}{logical. If TRUE the genotype by environment means is calculated.}

\item{colGen}{colour for genotype attributes on biplot. Defaults to "gray"}

\item{colEnv}{colour for environment attributes on biplot. Defaults to "darkred"}

\item{colSegment}{colour for segment or circle lines. Defaults to "gray"}

\item{colHull}{colour for hull when type=6. Defaults to "gray"}

\item{sizeGen}{text size for genotype labels. Defaults to 4}

\item{sizeEnv}{text size for environment labels. Defaults to 4}

\item{largeSize}{text size to use for larger labels where type=5, used for the}

\item{titles}{logical. If TRUE then include automatically generated titles}

\item{footnote}{logical. If TRUE then include automatically generated footbote
two selected genotypes, and where type=6, used for the outermost genotypes.
Defaults to 4.5}
}
\value{
A biplot of class  \code{ggplot}
}
\description{
Produces the AMMI biplot as an object of class 'ggplot'. It is possible to customize
it so that the stylistic attributes are to the user's liking.
}
\references{
Rodrigues PC, Monteiro A and Lourenco VM (2015). \emph{A robust AMMI model for the analysis of
genotype-by-environment data}. Bioinformatics, 32, 58–66.
}
