% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rAMMI.R
\name{rAMMI}
\alias{rAMMI}
\title{AMMI biplots with \pkg{ggplot2}}
\usage{
rAMMI(
  Data,
  genotype = "gen",
  environment = "env",
  response = "Y",
  rep = NULL,
  Ncomp = 2,
  type = "AMMI",
  colGen = "gray47",
  colEnv = "darkred",
  colSegment = "gray30",
  colHull = "gray30",
  sizeGen = 4,
  sizeEnv = 4,
  largeSize = 4.5,
  titles = TRUE,
  footnote = TRUE,
  axis_expand = 1.2,
  limits = TRUE,
  axes = TRUE,
  axislabels = TRUE
)
}
\arguments{
\item{Data}{a data frame}

\item{genotype}{name of the column that contains the genotypes}

\item{environment}{name of the column that contains the environments}

\item{response}{name of the column that contains the response}

\item{rep}{name of the column that contains the replications.If this argument
is NULL, there is no replications in the data.}

\item{Ncomp}{number of principal components that will be used in the analysis}

\item{type}{method. Either "AMMI", "rAMMI", "hAMMI", "gAMMI", "lAMMI" or
"ppAMMI".}

\item{colGen}{colour for genotype attributes on biplot. Defaults to "gray"}

\item{colEnv}{colour for environment attributes on biplot. Defaults to
"darkred"}

\item{colSegment}{colour for segment or circle lines. Defaults to "gray"}

\item{colHull}{colour for hull when type=6. Defaults to "gray"}

\item{sizeGen}{text size for genotype labels. Defaults to 4}

\item{sizeEnv}{text size for environment labels. Defaults to 4}

\item{largeSize}{text size to use for larger labels where type=5, used for the
two selected genotypes, and where type="Which won where/what", used for the
outermost genotypes. Defaults to 4.5}

\item{titles}{logical. If TRUE then include automatically generated titles}

\item{footnote}{logical. If TRUE then include automatically generated footbote
two selected genotypes, and where type=6, used for the outermost genotypes.}

\item{axis_expand}{multiplication factor to expand the axis limits by to
enable fitting of labels. Defaults to 1.2}

\item{limits}{logical. If TRUE then automatically rescale axes}

\item{axes}{logical. If TRUE then include x and y axes going through the
origin}

\item{axislabels}{logical. If TRUE then include automatically generated labels
for axes}
}
\value{
A biplot of class  \code{ggplot}
}
\description{
Produces the AMMI biplot as an object of class 'ggplot'. It is possible to
customize it so that the stylistic attributes are to the user's liking.
}
\examples{

library(geneticae)
# Data without replication
data(yan.winterwheat)
GGE1 <- rAMMI(yan.winterwheat, genotype="gen",environment="env", response="yield",
             type = "AMMI")
GGE1

# Data with replication
data(plrv)
GGE2 <- rAMMI(plrv, genotype="Genotype",environment="Locality", response="Yield",
             rep="Rep", type = "AMMI")
GGE2


}
\references{
Rodrigues PC, Monteiro A and Lourenco VM (2015). \emph{A robust
 AMMI model for the analysis of genotype-by-environment data}.
 Bioinformatics, 32, 58–66.
}
