% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rAMMIModel.R
\name{rAMMIModel}
\alias{rAMMIModel}
\title{Robust AMMI Model}
\usage{
rAMMIModel(
  Data,
  genotype = "gen",
  environment = "env",
  response = "Y",
  rep = NULL,
  Ncomp = 2,
  type = "AMMI"
)
}
\arguments{
\item{Data}{a dataframe with genotypes, environments and the phenotypic trait.}

\item{genotype}{column name containing genotypes.}

\item{environment}{column name containing environments.}

\item{response}{column name containing the phenotypic trait.}

\item{rep}{column name containing replications. If provided, means are calculated.}

\item{Ncomp}{number of principal components to retain.}

\item{type}{method for fitting: `"AMMI"`, `"rAMMI"`, `"hAMMI"`, `"gAMMI"`, `"lAMMI"` or `"ppAMMI"`.}
}
\description{
Fits a classical or robust Additive Main effects and Multiplicative Interaction (AMMI) model.
}
