% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGabriel.R
\name{WGabriel}
\alias{WGabriel}
\title{Weighted Gabriel imputation}
\usage{
WGabriel(DBmiss, Winf, Wsup)
}
\arguments{
\item{DBmiss}{a data frame or matrix that contains the genotypes in the rows and the
environments in the columns when there are no replications of the experiment.}

\item{Winf}{peso inferior}

\item{Wsup}{peso superior}
}
\description{
Weighted Gabriel imputation
}
\references{
FALTA
}
