% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{yan.winterwheat}
\alias{yan.winterwheat}
\title{Multi-environment trial of winter wheat in Ontario}
\format{A data frame with 162 observations on the following 3 variables.
gen genotype
env environment
yield yield in metric tons per hectare}
\usage{
data(yan.winterwheat)
}
\description{
Yield of 18 varieties of winter wheat grown at 9 environments in Ontario in 1993.
This data.frame was obtained from agridat Package.
}
\examples{
library(geneticae)
data(yan.winterwheat)
str(yan.winterwheat)
}
\references{
Weikai Yan and Manjit S. Kang and Baoluo Ma and Sheila Woods, 2007, GGE Biplot vs. AMMI
Analysis of Genotype-by-Environment Data, Crop Science, 2007, 47, 641–653. \url{http://doi.org/10.2135/cropsci2006.06.0374}
}
\keyword{datasets}
